sentinelle/requirements.txt
# =============================================================================
# requirements.txt -- Dependances Python du firmware de la sentinelle DTN
# =============================================================================
# Installation : pip install -r requirements.txt

# --- Cryptographie (obligatoire sur toutes les plateformes) ---
pycryptodome>=3.20.0

# --- Capteurs (Raspberry Pi uniquement -- decommenter sur Pi) ---
# adafruit-circuitpython-dht>=4.0.0
# RPi.bme280>=0.2.4
# smbus2>=0.4.3
# pyserial>=3.5

# --- QR code (optionnel, pour generation au deploiement) ---
qrcode[pil]>=7.4

# --- BLE (Raspberry Pi uniquement, pre-installe sur Pi OS) ---
# dbus-python>=1.3.2
# PyGObject>=3.42.0

# --- Tests ---
pytest>=8.0.0
